Avoid coverage bleed for AA arcs with tiny sweep angles
Because we outset the planes by 1/2 pixel, the intersection of those
planes could extend well past the center of the circle. To fix this (and
cut down on fill-rate), do a geometric clip of the vertices against the
perpendicular diameter of the circle (while still leaving a half-pixel
for the coverage to fade out).

Bug: b/112018278
Change-Id: Icfb38ecfe57aad823c49fc4cadd0538a37dc1b7b
Reviewed-on: https://skia-review.googlesource.com/145326
Reviewed-by: Brian Salomon <bsalomon@google.com>
Commit-Queue: Brian Osman <brianosman@google.com>
(cherry-picked from 9a24fee9a958142837fc2fc8a545882d016d7c0d)

Signed-off-by: Joe Maples <joe@frap129.org>